\name{gSNR-package}
\alias{gSNR-package}
\alias{gSNR}
\docType{package}
NA
\title{
\packageTitle{gSNR}
}
\description{
\packageDescription{gSNR}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{gSNR}
\packageIndices{gSNR}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{gSNR}

Maintainer: \packageMaintainer{gSNR}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
